feat(slatetohtml): convert line breaks to br tags
