+++ src/serializers/blocks.ts
@@ -1,0 +1,66 @@
+// See: <https://html.spec.whatwg.org/#the-css-user-agent-style-sheet-and-presentational-hints>
+// From: https://www.npmjs.com/package/rehype-minify-whitespace
+const blocks = [
+  'address', // Flow content.
+  'article', // Sections and headings.
+  'aside', // Sections and headings.
+  'blockquote', // Flow content.
+  'body', // Page.
+  'br', // Contribute whitespace intrinsically.
+  'caption', // Similar to block.
+  'center', // Flow content, legacy.
+  'col', // Similar to block.
+  'colgroup', // Similar to block.
+  'dd', // Lists.
+  'dialog', // Flow content.
+  'dir', // Lists, legacy.
+  'div', // Flow content.
+  'dl', // Lists.
+  'dt', // Lists.
+  'figcaption', // Flow content.
+  'figure', // Flow content.
+  'footer', // Flow content.
+  'form', // Flow content.
+  'h1', // Sections and headings.
+  'h2', // Sections and headings.
+  'h3', // Sections and headings.
+  'h4', // Sections and headings.
+  'h5', // Sections and headings.
+  'h6', // Sections and headings.
+  'head', // Page.
+  'header', // Flow content.
+  'hgroup', // Sections and headings.
+  'hr', // Flow content.
+  'html', // Page.
+  'legend', // Flow content.
+  'li', // Block-like.
+  'li', // Similar to block.
+  'listing', // Flow content, legacy
+  'main', // Flow content.
+  'menu', // Lists.
+  'nav', // Sections and headings.
+  'ol', // Lists.
+  'optgroup', // Similar to block.
+  'option', // Similar to block.
+  'p', // Flow content.
+  'plaintext', // Flow content, legacy
+  'pre', // Flow content.
+  'section', // Sections and headings.
+  'summary', // Similar to block.
+  'table', // Similar to block.
+  'tbody', // Similar to block.
+  'td', // Block-like.
+  'td', // Similar to block.
+  'tfoot', // Similar to block.
+  'th', // Block-like.
+  'th', // Similar to block.
+  'thead', // Similar to block.
+  'tr', // Similar to block.
+  'ul', // Lists.
+  'wbr', // Contribute whitespace intrinsically.
+  'xmp', // Flow content, legacy
+]
+
+export const isBlock = (tagName: string) => {
+  return blocks.includes(tagName)
+}
